requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# NLP
transformers>=4.35.0
torch>=2.0.0
nltk>=3.8.0
rapidfuzz>=3.0.0

# ML/Quantization
requests>=2.31.0
scikit-learn>=1.3.0


# Streamlit
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
scipy>=1.11.0